feat: change project-structure / vite-config
