+++ src/index.js
@@ -1,5 +1,5 @@
-import './index.scss';
-import App from './App.js';
+import 'styles/index.scss';
+import App from 'ui/App.js';
 
 document.addEventListener('DOMContentLoaded', () => {
   const app = new App();
