+++ src/ui/pages/index.js
@@ -1,0 +1,5 @@
+import Login from 'pages/Login';
+
+export {
+  Login,
+};
